refactor: refactored ecogauge to MVP format
