+++ app/src/main/res/layout/fragment_eco_gauge.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:padding="15dp"
-    tools:context=".view.EcoGaugeFragment">
+    tools:context=".view.ecogauge.EcoGaugeFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
